Move dist_string_from_package_record to .utils
